+++ +gams/+transfer/+symbol/Equation.m
@@ -131,7 +131,7 @@ classdef Equation < gams.transfer.symbol.Abstract
 
     methods (Hidden, Access = {?gams.transfer.symbol.Abstract, ?gams.transfer.Container})
 
-        function obj = Equation(container, name, type, init_domains, init_records)
+        function obj = Equation(container, name, type, init_records)
             obj.container_ = container;
             obj.name_ = name;
             obj.def_ = gams.transfer.symbol.definition.Equation.construct(type);
@@ -204,7 +204,7 @@ classdef Equation < gams.transfer.symbol.Abstract
                     if strcmpi(varargin{index}, 'description')
                         index = index + 1;
                         gams.transfer.utils.Validator.minargin(nargin, index);
-                        description = gams.transfer.utils.Validator('name', index, varargin{index}) ...
+                        description = gams.transfer.utils.Validator('description', index, varargin{index}) ...
                             .symbolDescription().value;
                         has_description = true;
                         index = index + 1;
@@ -244,7 +244,7 @@ classdef Equation < gams.transfer.symbol.Abstract
                 error(e.message);
             end
 
-            obj = gams.transfer.symbol.Equation(container, name, type, ~has_domains && ~has_size, ~has_records);
+            obj = gams.transfer.symbol.Equation(container, name, type, ~has_records);
             if has_description
                 obj.description_ = description;
             end
@@ -325,12 +325,12 @@ classdef Equation < gams.transfer.symbol.Abstract
                     destination.removeSymbols(obj.name_);
                     symbol = destination.addEquation(obj.name_, obj.def_.type);
                 end
-                symbol.copyFrom(obj);
-                symbol.def.switchContainer(destination);
             else
                 symbol = destination.addEquation(obj.name_, obj.def_.type);
-                symbol.copyFrom(obj);
             end
+
+            symbol.copyFrom(obj);
+            symbol.def.switchContainer(destination);
         end
 
     end