+++ +gams/+transfer/+symbol/Parameter.m
@@ -81,7 +81,7 @@ classdef Parameter < gams.transfer.symbol.Abstract
 
     methods (Hidden, Access = {?gams.transfer.symbol.Abstract, ?gams.transfer.Container})
 
-        function obj = Parameter(container, name, init_domains, init_records)
+        function obj = Parameter(container, name, init_records)
             obj.container_ = container;
             obj.name_ = name;
             obj.def_ = gams.transfer.symbol.definition.Parameter();
@@ -149,7 +149,7 @@ classdef Parameter < gams.transfer.symbol.Abstract
                     if strcmpi(varargin{index}, 'description')
                         index = index + 1;
                         gams.transfer.utils.Validator.minargin(nargin, index);
-                        description = gams.transfer.utils.Validator('name', index, varargin{index}) ...
+                        description = gams.transfer.utils.Validator('description', index, varargin{index}) ...
                             .symbolDescription().value;
                         has_description = true;
                         index = index + 1;
@@ -189,7 +189,7 @@ classdef Parameter < gams.transfer.symbol.Abstract
                 error(e.message);
             end
 
-            obj = gams.transfer.symbol.Parameter(container, name, ~has_domains && ~has_size, ~has_records);
+            obj = gams.transfer.symbol.Parameter(container, name, ~has_records);
             if has_description
                 obj.description_ = description;
             end
@@ -284,12 +284,12 @@ classdef Parameter < gams.transfer.symbol.Abstract
                     destination.removeSymbols(obj.name_);
                     symbol = destination.addParameter(obj.name_);
                 end
-                symbol.copyFrom(obj);
-                symbol.def.switchContainer(destination);
             else
                 symbol = destination.addParameter(obj.name_);
-                symbol.copyFrom(obj);
             end
+
+            symbol.copyFrom(obj);
+            symbol.def.switchContainer(destination);
         end
 
     end